python-workspace/requirements.txt
# Runtime dependencies
strands-agents
strands-agents-tools
litellm
python-dotenv

# Testing
pytest
